+++ CHANGELOG.md
@@ -37,8 +37,11 @@
 - Move adding health ids from edit patient behind a feature flag
 - Display National Health ID before BP Passports in `EditPatientScreen`
 - Bump google-services to 4.3.10
-- [In Progress: 11 Jan 2022] Implement adding NHID in `PatientEditScreen`
+- Add highlight when a new alternate id is added 
 - [In Progress: 17 Jan 2022] Implement adding scan support for BpPassport in `EditPatientScreen`
+
+### Changes
+- Implement adding scan support for India NHID in `EditPatientScreen`
 
 ## On Demo